Refactor and make the table independent
